debt - getActiveWindowId is electron only
